backend/requirements.txt
aws-cdk-lib==2.248.0
constructs==10.6.0
boto3==1.42.83
cdk-nag==2.37.55
python-dotenv==1.1.0
